move homepage to app router
